+++ GCOPY/src-java/br/com/dnolive/services/ServiceContrato.java
@@ -4,7 +4,7 @@ import br.com.dnolive.controllers.DaoFactory;
 import br.com.dnolive.core.ImplementsService;
 import br.com.dnolive.models.Contrato;
 
-public class ServiceContrato extends ImplementsService<Contrato, Long> 
+public class ServiceContrato extends ImplementsService<Contrato, Integer> 
 {
 	public ServiceContrato() {
 		super(DaoFactory.DAO.CONTRATO);
