+++ GCOPY/src-java/br/com/dnolive/services/ServiceEquipamento.java
@@ -4,7 +4,7 @@ import br.com.dnolive.controllers.DaoFactory;
 import br.com.dnolive.core.ImplementsService;
 import br.com.dnolive.models.Equipamento;
 
-public class ServiceEquipamento extends ImplementsService<Equipamento, Long>
+public class ServiceEquipamento extends ImplementsService<Equipamento, Integer>
 {
 	public ServiceEquipamento() {
 		super(DaoFactory.DAO.EQUIPAMENTO);
