+++ GCOPY/src-java/br/com/dnolive/services/ServiceLocadora.java
@@ -4,7 +4,7 @@ import br.com.dnolive.controllers.DaoFactory;
 import br.com.dnolive.core.ImplementsService;
 import br.com.dnolive.models.Locadora;
 
-public class ServiceLocadora extends ImplementsService<Locadora, Long> 
+public class ServiceLocadora extends ImplementsService<Locadora, Integer> 
 {
 	public ServiceLocadora() {
 		super(DaoFactory.DAO.LOCADORA);
