+++ GCOPY/src-java/br/com/dnolive/services/ServiceMovimento.java
@@ -4,9 +4,10 @@ import br.com.dnolive.controllers.DaoFactory;
 import br.com.dnolive.core.ImplementsService;
 import br.com.dnolive.models.Movimento;
 
-public class ServiceMovimento extends ImplementsService<Movimento, Long> 
+public class ServiceMovimento extends ImplementsService<Movimento, Integer> 
 {
 	public ServiceMovimento() {
 		super(DaoFactory.DAO.MOVIMENTO);
 	}
+	
 }
