+++ GCOPY/src-java/br/com/dnolive/services/ServiceMovimitem.java
@@ -4,7 +4,7 @@ import br.com.dnolive.controllers.DaoFactory;
 import br.com.dnolive.core.ImplementsService;
 import br.com.dnolive.models.Movimitem;
 
-public class ServiceMovimitem extends ImplementsService<Movimitem, Long> 
+public class ServiceMovimitem extends ImplementsService<Movimitem, Integer> 
 {
 	public ServiceMovimitem() {
 		super(DaoFactory.DAO.MOVIMITEM);
