+++ GCOPY/src-java/br/com/dnolive/services/ServiceSetor.java
@@ -4,7 +4,7 @@ import br.com.dnolive.controllers.DaoFactory;
 import br.com.dnolive.core.ImplementsService;
 import br.com.dnolive.models.Setor;
 
-public class ServiceSetor extends ImplementsService<Setor, Long> 
+public class ServiceSetor extends ImplementsService<Setor, Integer> 
 {
 	public ServiceSetor() {
 		super(DaoFactory.DAO.SETOR);
